[REFACTOR] #138: components 폴더구조 변경
