+++ frontend/src/containers/OptionPage/OptionSelectContainer/DefaultOptionContainer.tsx
@@ -4,7 +4,7 @@ import { BodyKrRegular4, HeadingEn4, HeadingKrMedium7 } from '../../../styles/ty
 import { useState } from 'react';
 import RoundButton from '../../../components/common/buttons/RoundButton';
 import { CheckIcon } from '../../../components/common/icons/Icons';
-import HmgTag from '../../../components/hmgTag/HmgTag';
+import HmgTag from '../../../components/common/hmgTag/HmgTag';
 
 export default function DefaultOptionContainer() {
   const [selectedOptions, setSelectedOptions] = useState<number[]>([]);
