+++ frontend/src/containers/OptionPage/OptionSelectContainer/ExtraOptionContainer copy.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { styled } from 'styled-components';
 import RoundButton from '../../../components/common/buttons/RoundButton';
 import { CheckIcon } from '../../../components/common/icons/Icons';
-import HmgTag from '../../../components/hmgTag/HmgTag';
+import HmgTag from '../../../components/common/hmgTag/HmgTag';
 import { BodyKrRegular4, HeadingEn4, HeadingKrMedium7 } from '../../../styles/typefaces';
 import DefaultCardStyle from '../../../components/common/card/DefaultCardStyle';
